fix(date-picker): missing last day of week offset calculation for calendars using firstDayOfWeek (#426)
* fix(date-picker): date adapters getDay() and create() do not behave consistent

This will change the default for `NgpNativeDateAdapter.getDay()` to be a 1-7 number with 1=Monday - 7=Sunday instead of a 0-6 number with 0=Sunday. It now aligns with the `NgpLuxonDateAdapter` and the previously introduced firstDayOfWeek change.

Also since we now do not have a 0 based day index, the month in the `NgpNativeDateAdapter.create()` will now also align to the `NgpLuxonDateAdapter` by using 1-12 instead of 0-11. This resolves the different handling of the create in the date adapters.

Additionally, introduce a new testing module for date adapters to standardize tests across different date adapter implementations.

* feat(date-time): add NgpNativeDateAdapter export

* fix(date-picker): missing last day of week offset calculation for calendars using firstDayOfWeek

Adding a new method to calculate the last day of the week offset for the last week in the month in the NgpDatePickerRowRender.