fix: add compatibility for @angular/cdk@20 caused by removed deprecations in DomPortalOutlet (#486)
* fix: use new DomPortalOutlet constructor signature (ng@20+ breaking change)

DomPortalOutlet's constructor was changed in v20 after removing the
deprecated 2nd param `_unusedComponentFactoryResolver?: any`.
This causes several issues when using ng@20+ which are not
immediately apparent.

* fix: ensure compatibility with angular < v20

In ng > v20, `_unusedComponentFactoryResolver` and `_document`
constructor params for DomComponentOutlet have been removes. This fix
make a runtime check to ensure the injector is in the right position

* fix: replace the flaky fix with a reliable CDK version check

- improve comments to detail the need for the change, with references